deprecate TICK_INTERVAL_SECONDS in favor of TICK_INTERVAL_MILLIS, closes #609 (#610)
* 1 - deprecate TICK_INTERVAL_SECONDS in favor of TICK_INTERVAL_MILLIS

* 2 - fix bug in deprecation printing logic

* 3 - fix bug during conversion process of seconds to millis

* 4 - fix bug in updateRefTime in trader#Start()

+++ plugins/metricsTracker.go
@@ -62,7 +62,7 @@ type commonProps struct {
 	GoVersion                        string  `json:"go_version"`
 	GuiVersion                       string  `json:"gui_version"`
 	Strategy                         string  `json:"strategy"`
-	UpdateTimeIntervalSeconds        int32   `json:"update_time_interval_seconds"`
+	UpdateTimeIntervalSeconds        float64 `json:"update_time_interval_seconds"`
 	Exchange                         string  `json:"exchange"`
 	TradingPair                      string  `json:"trading_pair"`
 	SecondsSinceStart                float64 `json:"seconds_since_start"`
@@ -155,7 +155,7 @@ func MakeMetricsTrackerCli(
 	goVersion string,
 	guiVersion string,
 	strategy string,
-	updateTimeIntervalSeconds int32,
+	updateTimeIntervalSeconds float64,
 	exchange string,
 	tradingPair string,
 	isDisabled bool,
